backend/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cryptography==38.0.4
cycler==0.11.0
daphne==4.0.0
decorator==5.1.1
Django==3.2.16
firebase-admin==6.0.1
fonttools==4.38.0
gcloud==0.17.0
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-aiplatform==1.19.1
google-cloud-bigquery==2.34.4
google-cloud-core==2.3.2
google-cloud-firestore==2.7.3
google-cloud-resource-manager==1.6.3
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
greenlet==2.0.1
grpc-google-iam-v1==0.12.4
grpcio==1.51.1
grpcio-status==1.48.2
h11==0.14.0
httplib2==0.21.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.1.0
incremental==22.10.0
jws==0.1.3
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.5.3
msgpack==1.0.4
numpy==1.21.6
oauth2client==3.0.0
obspy==1.3.1
packaging==21.3
Pillow==9.3.0
proto-plus==1.22.1
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2022.6
redis==4.4.0
requests==2.28.1
requests-toolbelt==0.7.0
rsa==4.9
scipy==1.7.3
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
uvicorn==0.20.0
zipp==3.11.0
zope.interface==5.5.2
